% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build2video.R
\name{build2video}
\alias{build2video}
\title{Build Video from Quotes}
\usage{
build2video(project_name)
}
\arguments{
\item{project_name}{Character string, the name of the project directory to use.}
}
\value{
Invisible NULL, creates the final video as a side effect.
}
\description{
This function compiles a video by generating images, voice-overs, individual videos,
and the final combined video using files from the specified project directory.
The final video is saved in the project directory.
}
\examples{
build2video("elon")
}
